chore: move util to internal package
