+++ pkg/golinters/gochecknoinits.go
@@ -9,6 +9,7 @@ import (
 	"golang.org/x/tools/go/analysis"
 
 	"github.com/golangci/golangci-lint/pkg/golinters/goanalysis"
+	"github.com/golangci/golangci-lint/pkg/golinters/internal"
 	"github.com/golangci/golangci-lint/pkg/lint/linter"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
@@ -64,7 +65,7 @@ func checkFileForInits(f *ast.File, fset *token.FileSet) []result.Issue {
 		if name == "init" && funcDecl.Recv.NumFields() == 0 {
 			res = append(res, result.Issue{
 				Pos:        fset.Position(funcDecl.Pos()),
-				Text:       fmt.Sprintf("don't use %s function", formatCode(name, nil)),
+				Text:       fmt.Sprintf("don't use %s function", internal.FormatCode(name, nil)),
 				FromLinter: gochecknoinitsName,
 			})
 		}
