+++ pkg/golinters/gocyclo.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/golangci/golangci-lint/pkg/config"
 	"github.com/golangci/golangci-lint/pkg/golinters/goanalysis"
+	"github.com/golangci/golangci-lint/pkg/golinters/internal"
 	"github.com/golangci/golangci-lint/pkg/lint/linter"
 	"github.com/golangci/golangci-lint/pkg/result"
 )
@@ -62,7 +63,7 @@ func runGoCyclo(pass *analysis.Pass, settings *config.GoCycloSettings) []goanaly
 
 	for _, s := range stats {
 		text := fmt.Sprintf("cyclomatic complexity %d of func %s is high (> %d)",
-			s.Complexity, formatCode(s.FuncName, nil), settings.MinComplexity)
+			s.Complexity, internal.FormatCode(s.FuncName, nil), settings.MinComplexity)
 
 		issues = append(issues, goanalysis.NewIssue(&result.Issue{
 			Pos:        s.Pos,
